Alarm not marked as clonable, since breaking clone() contract
FindBugs note:
This non-final class defines a clone() method that does not call
super.clone(). If this class ("A") is extended by a subclass ("B"), and
the subclass B calls super.clone(), then it is likely that B's clone()
method will return an object of type A, which violates the standard
contract for clone().