Redirect the user from MainActivity to AlarmActivity if an alarm is ringing
A reference to the ringing alarm is stored in SFApplication. This might be
temporary, in case we find a better way to store the state. Only when the alarm
is actually ringing will the user not be able to access MainActivity and
toggle the alarm off, it's still possible while snoozing.